Started implementation of Cauchy directions

+++ src/main/lp/sleqp_lpi.h
@@ -18,8 +18,8 @@ extern "C" {
                                            SLEQP_LPI_FREE free_problem);
 
   SLEQP_RETCODE sleqp_lpi_create_problem(SleqpLPi* lp_interface,
-                                         int num_variables,
-                                         int num_constraints);
+                                         size_t num_variables,
+                                         size_t num_constraints);
 
   SLEQP_RETCODE sleqp_lpi_solve(SleqpLPi* lp_interface,
                                 double* objective,
@@ -30,12 +30,12 @@ extern "C" {
                                 double* vars_ub);
 
   SLEQP_RETCODE sleqp_lpi_get_solution(SleqpLPi* lp_interface,
-                                       int num_variables,
+                                       size_t num_variables,
                                        double* objective_value,
                                        double* solution_values);
 
   SLEQP_RETCODE sleqp_lpi_get_varstats(SleqpLPi* lp_interface,
-                                       int num_variables,
+                                       size_t num_variables,
                                        SLEQP_BASESTAT* variable_stats);
 
   SLEQP_RETCODE sleqp_lpi_free(SleqpLPi** lp_interface);
